+++ example/lib/utils/Constants.dart
@@ -1,9 +1,9 @@
 //   **** Routes *****
-const CONTAINERS_ROUTE = '/containers';
-const TEXTS_ROUTE = '/texts';
-const BUTTONS_ROUTE = '/buttons';
-const TEXT_INPUTS_ROUTE = '/textInputs';
-const DATE_PICKERS_ROUTE = '/datePicker';
+const String routeContainers = '/containers';
+const String routeTexts = '/texts';
+const String routeButtons = '/buttons';
+const String routeTextInputs = '/textInputs';
+const String routeDatePickers = '/datePicker';
 
 //   **** Colors *****
-const PRIMARY_COLOR = 0xFF9DCCD1;
+const int colorPrimary = 0xFF9DCCD1;