+++ lib/models/index.dart
@@ -1,5 +1,5 @@
+export './CommonButtonModel.dart';
 export './CommonContainerModel.dart';
-export './CommonTextModel.dart';
 export './CommonIcon.dart';
 export './CommonTextInputModel.dart';
-export './CommonButtonModel.dart';
+export 'common_text_model.dart';