+++ lib/utils/index.dart
@@ -1,3 +1,4 @@
 export './Constants.dart';
 export './Helper.dart';
 export 'i18n_local_types.dart';
+export 'log.dart';